v3

+++ src/main/java/com/jesuscast/reposteriaanita/models/Pastel.java
@@ -5,7 +5,7 @@ public class Pastel extends Producto {
     private String relleno;
     private String desing;
     public Pastel(){}
-    public Pastel(String nombreProducto, int cantidad, int precio, int cantidadPersonas, String relleno, String desing) {
+    public Pastel(String nombreProducto, int cantidad, double precio, int cantidadPersonas, String relleno, String desing) {
         super(nombreProducto, cantidad, precio);
         this.cantidadPersonas = cantidadPersonas;
         this.relleno = relleno;
